soc: qcom: glink_probe: Add support for spi transport
Add support to probe the glink spi transport for the wdsp processor.
The WDSP processor notifiers about SSR through a component manager
instead of the PIL SSR framework. If the SPI transport is required,
register the SPI transport immediately instead of waiting for an up
notification from the SSR framework.

Change-Id: I313b6acae305d43a0e328c78fde255243be76eab
Signed-off-by: Chris Lew <clew@codeaurora.org>
Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>
Signed-off-by: Helium-Studio <67852324+Helium-Studio@users.noreply.github.com>