power: supply: qcom: Fix NULL pointer de-reference due late IIO init
In the current flow, there can be a race condition between
the IIO channel initialization and the access to the IIO channels.

To elaborate, in current flow:
      1. power_supply notifier registration
      2. profile_load_work scheduled
      3. IIO initialization

Here #1 and #2, both can lead to 'state_change_work()' to get called
before the #3(IIO init), and we will end up seeing NULL pointer
dereference issue, as the IIO channels will get accessed before
being initialized.

Fix this by changing the flow to:
       1. IIO initialization
       2. power_supply notifier registration
       3. profile_load_work scheduled

This ensures that IIO channels are initialized early before
a possible access is made.

Change-Id: I5ee41e5fc41b318ba5f3c358865c4c6c74f88453
Signed-off-by: Rakesh Kota <quic_kotarake@quicinc.com>
Signed-off-by: Helium-Studio <67852324+Helium-Studio@users.noreply.github.com>